Move `indexDocuments` into elastic.go
