+++ elastic.go
@@ -3,13 +3,16 @@ package main
 import (
 	"bytes"
 	"context"
+	"crypto/rand"
 	"crypto/tls"
 	"crypto/x509"
+	"encoding/base64"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"net/http"
 	"reflect"
+	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -328,4 +331,42 @@ func restoreDocuments(sb s3Backend, c elastic.Client, keyPath, indexName string)
 	}
 
 	return err
+}
+
+func (es *esClient) indexDocuments(indexName string) {
+	log.Println("Indexing the documents...")
+	for i := 1; i <= 100; i++ {
+		str := fmt.Sprintf(`{"%s" : "%s"}`, generateRandomBytes(20), generateRandomBytes(10))
+		res, err := es.client.Index(
+			indexName,
+			strings.NewReader(str),
+			es.client.Index.WithDocumentID(strconv.Itoa(i)),
+		)
+
+		if err != nil || res.IsError() {
+			log.Error("Error happens here")
+			log.Fatalf("Error: %s: %s", err, res)
+		}
+		time.Sleep(time.Millisecond * 50)
+	}
+}
+
+// GenerateRandomBytes generates a rnd string
+func generateRandomBytes(n int) string {
+	b := make([]byte, n)
+	_, err := rand.Read(b)
+	if err != nil {
+		log.Fatal(err)
+	}
+	rb := base64.StdEncoding.EncodeToString(b)
+
+	return rb
+}
+
+func esURI (c elasticConfig) string {
+	URI := c.host
+
+	URI = fmt.Sprintf(URI+":%d", c.port)
+
+	return URI
 }
